chore(bankingBridge): Move connector to ee
