+++ .codecov.yml
@@ -14,7 +14,7 @@ ignore:
   - "internal/connectors/engine/activities/workflow_run_generated.go"
   - "internal/connectors/plugins/public/adyen/client/client_generated.go"
   - "internal/connectors/plugins/public/atlar/client/client_generated.go"
-  - "internal/connectors/plugins/public/bankingbridge/client/client_generated.go"
+  - "ee/plugins/bankingbridge/client/client_generated.go"
   - "internal/connectors/plugins/public/bankingcircle/client/client_generated.go"
   - "internal/connectors/plugins/public/currencycloud/client/client_generated.go"
   - "internal/connectors/plugins/public/mangopay/client/client_generated.go"
