+++ ee/plugins/bankingbridge/balances.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"math/big"
 
-	"github.com/formancehq/payments/internal/connectors/plugins/public/bankingbridge/client"
+	"github.com/formancehq/payments/ee/plugins/bankingbridge/client"
 	"github.com/formancehq/payments/internal/models"
 )
