+++ ee/plugins/bankingbridge/balances_test.go
@@ -6,8 +6,8 @@ import (
 	"errors"
 	"time"
 
+	"github.com/formancehq/payments/ee/plugins/bankingbridge/client"
 	"github.com/formancehq/payments/internal/connectors/plugins"
-	"github.com/formancehq/payments/internal/connectors/plugins/public/bankingbridge/client"
 	"github.com/formancehq/payments/internal/models"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
