+++ ee/plugins/bankingbridge/capabilities.go
