+++ ee/plugins/bankingbridge/client/client.go
