+++ ee/plugins/bankingbridge/client/client_generated.go
