+++ ee/plugins/bankingbridge/client/util.go
