+++ ee/plugins/bankingbridge/client/util_test.go
@@ -4,7 +4,7 @@ import (
 	"net/url"
 	"testing"
 
-	"github.com/formancehq/payments/internal/connectors/plugins/public/bankingbridge/client"
+	"github.com/formancehq/payments/ee/plugins/bankingbridge/client"
 )
 
 func TestRawQuery(t *testing.T) {
