+++ ee/plugins/bankingbridge/config.go
