+++ ee/plugins/bankingbridge/payments.go
@@ -7,7 +7,7 @@ import (
 	"math/big"
 
 	"github.com/formancehq/go-libs/v3/pointer"
-	"github.com/formancehq/payments/internal/connectors/plugins/public/bankingbridge/client"
+	"github.com/formancehq/payments/ee/plugins/bankingbridge/client"
 	"github.com/formancehq/payments/internal/models"
 )
