+++ internal/connectors/plugins/public/list.go
@@ -1,22 +1,21 @@
 package public
 
 import (
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/adyen"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/atlar"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/bankingbridge"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/bankingcircle"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/column"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/currencycloud"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/dummypay"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/generic"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/increase"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/mangopay"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/modulr"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/moneycorp"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/plaid"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/powens"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/qonto"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/stripe"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/tink"
-    _ "github.com/formancehq/payments/internal/connectors/plugins/public/wise"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/adyen"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/atlar"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/bankingcircle"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/column"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/currencycloud"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/dummypay"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/generic"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/increase"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/mangopay"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/modulr"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/moneycorp"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/plaid"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/powens"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/qonto"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/stripe"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/tink"
+	_ "github.com/formancehq/payments/internal/connectors/plugins/public/wise"
 )
