+++ internal/connectors/plugins/registry/enterprise_ce.go
@@ -6,6 +6,7 @@ package registry
 // Enterprise Edition. When a user attempts to use one of these connectors
 // on a Community Edition binary, we can return a more helpful error message.
 var EnterpriseOnlyPlugins = map[string]struct{}{
+	"bankingbridge": {},
 	"fireblocks":    {},
 	"coinbaseprime": {},
 }
